fix(a11y): dialog close button aria-label
