+++ packages/i18n/src/sv.json
@@ -153,6 +153,8 @@
   "scheduled departure { time } estimated { estimate }": "Planerad avgång kl. { time } {estimate, select, undefined {} other {(beräknad {estimate})}}",
   "scheduled arrival { time } estimated { estimate }": "Planerad ankomst kl. { time } {estimate, select, undefined {} other {(beräknad {estimate})}}",
 
+  "close dialog": "Stäng dialogrutan",
+
   "stations": {
     "TPE": "Tammerfors",
     "PAU": "Böle biltågsstation",
